Adding saucelabs login test
